Move `UiImage` from `ui_node` to the `widget::image` module (#16084)
# Objective

`UiImage` isn't just a general image component now, it's the defining
component for the image widget so it belongs in the image widget's
module.
